updated usecase

+++ RUSAL.MetalTapping.BLL/Application/Contracts/OrderRequest.cs
@@ -1,4 +1,4 @@
 ﻿namespace RUSAL.MetalTapping.BLL.Application.Contracts
 {
-    public record OrderRequest(string metalMarkName, decimal requiredMetalWeight);
+    public record OrderRequest(string metalMarkName, double requiredMetalWeight);
 }
