+++ RUSAL.MetalTapping.BLL/Application/DTOs/BuildingMetalInfo.cs
@@ -5,7 +5,7 @@
         public Guid BuildingId { get; set; }
         public Guid MetalMarkId { get; set; }
         public int PotsCount { get; set; }
-        public decimal TotalMetalWeight { get; set; }
+        public double TotalMetalWeight { get; set; }
         public List<PotDto> Pots { get; set; }
     }
 }
